Add files via upload
A gas turbine is a type of engine that converts the energy of a fuel into mechanical power. The dataset includes 11 variables that measure the ambient conditions (such as temperature, pressure and humidity), the performance of the turbine (such as inlet and outlet temperatures, exhaust pressure and energy yield), and the emissions of the turbine (such as carbon monoxide and nitrogen oxides). The problem statement is to predict the turbine energy yield (TEY), which is the amount of electrical energy produced by the turbine, using the ambient variables as features. Features are the input variables that are used to train a machine learning model. The goal is to find a model that can accurately estimate the TEY based on the ambient conditions